requirements.txt
flask
flask-cors
torch
torchvision
opencv-python-headless
pillow
numpy